if user is not authenticated and navigates to dashboard by url there wont be nothing to see ot do

+++ src/components/MeetingsTable.jsx
@@ -9,53 +9,58 @@ import axios from "axios";
 const MeetingsTable = () => {
   const { scheduledMeetings, setMeetings } = useMeetingContext();
   const { uppdateMeeting, setUppdateMeeting } = useMeetingContext();
+  const { logged, setLogged } = useMeetingContext();
 
   const deleteMeeting = async (id) => {
     //let filtered = scheduledMeetings.filter(element => element.id !== id) //Used when running local meeting list
     //setMeetings(filtered)
 
-    try {
-      const response = await axios.delete(
-        `http://localhost:8080/api/v1/project/meetings/delete?id=${id}`,
-        {
-          method: "DELETE",
-          mode: "cors",
-          auth: {
-            username: localStorage.getItem("username"),
-            password: localStorage.getItem("password"),
-          },
-        }
-      );
-      if (response.status === 200) {
-        axios
-          .get("http://localhost:8080/api/v1/project/meetings", {
+    if (logged) {
+      try {
+        const response = await axios.delete(
+          `http://localhost:8080/api/v1/project/meetings/delete?id=${id}`,
+          {
+            method: "DELETE",
+            mode: "cors",
             auth: {
               username: localStorage.getItem("username"),
               password: localStorage.getItem("password"),
             },
-          })
-          .then((res) => {
-            setMeetings(res.data);
-          });
+          }
+        );
+        if (response.status === 200) {
+          axios
+            .get("http://localhost:8080/api/v1/project/meetings", {
+              auth: {
+                username: localStorage.getItem("username"),
+                password: localStorage.getItem("password"),
+              },
+            })
+            .then((res) => {
+              setMeetings(res.data);
+            });
+        }
+      } catch (error) {
+        console.log("Something went wrong when deleting meeting: " + error);
       }
-    } catch (error) {
-      console.log("Something went wrong when deleting meeting: " + error);
     }
   };
 
   function editMeeting(id) {
-    let meeting = scheduledMeetings.filter((element) => element.id === id);
-    console.log("meeting to be edited" + meeting);
-    setUppdateMeeting({
-      uppdate: true,
-      id: meeting[0].id,
-      title: meeting[0].title,
-      date: meeting[0].date,
-      time: meeting[0].time,
-      level: meeting[0].level,
-      participants: meeting[0].participants,
-      description: meeting[0].description,
-    });
+    if (logged) {
+      let meeting = scheduledMeetings.filter((element) => element.id === id);
+      console.log("meeting to be edited" + meeting);
+      setUppdateMeeting({
+        uppdate: true,
+        id: meeting[0].id,
+        title: meeting[0].title,
+        date: meeting[0].date,
+        time: meeting[0].time,
+        level: meeting[0].level,
+        participants: meeting[0].participants,
+        description: meeting[0].description,
+      });
+    }
   }
 
   return (
